requirements.txt
Flask
pymongo
pyyaml
requests
flake8
pylint
requests-mock
coverage
jsonschema
